microservices/hash-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
thrift==0.14.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-httpx==0.41b0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-otlp==1.20.0
pyyaml==6.0.1
pydantic-settings==2.1.0
prometheus_client==0.15.0